Added link to list group item
File title is clickable and triggering as separate event from clicking
anywhere else in the list group item. This could be useful for
triggering ui components when the line is selected vs. a different
action when the title itself is clicked.

Propegation of click events is stopped by the title when it is clicked
so a list group item click event is not also fired