Modularized view controller rendering
Created renderStatic and appendStatic functions that take a target
element and a template file to render. The static version of the
function assumes no data values are passed to the template.

Next Step: make data an optional third parameter to pass in and if
there is data, pass it through to the handlebars template.